# tanks_backend/src/main/resources/documentation/openAPI.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Tanks Online API",
    "description": "API documentation for ATA systems multiplayer game",
    "contact": {
      "name": "Antitalent Team",
      "email": "omazicantonio7@gmail.com"
    },
    "license": {
      "name": "MIT License",
      "url": "https://opensource.org/licenses/MIT"
    },
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Local development server"
    }
  ],
...
  "paths": {
...
    "/api/game-sessions/{gameSessionId}/start": {
      "put": {
        "tags": [
          "Game Sessions"
        ],
        "summary": "Start game session",
        "description": "Starts an existing game session by its ID",
        "operationId": "startGameSession",
        "parameters": [
          {
            "name": "gameSessionId",
            "in": "path",
            "description": "ID of the game session to start",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "400": {
            "description": "Game session cannot be started",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/GameSession"
                }
              }
            }
          },
          "404": {
            "description": "Game session not found",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/GameSession"
                }
              }
            }
          },
          "200": {
            "description": "Game session started successfully",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/GameSession"
                }
              }
            }
          }
        }
      }
    },
    "/api/game-sessions/{gameSessionId}/end": {
      "put": {
        "tags": [
          "Game Sessions"
        ],
        "summary": "End game session",
        "description": "Ends a running game session and records the results",
        "operationId": "endGameSession",
        "parameters": [
          {
            "name": "gameSessionId",
            "in": "path",
            "description": "ID of the game session to end",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GameSessionUpdateDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "400": {
            "description": "Game session cannot be ended",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/GameSession"
                }
              }
            }
          },
          "200": {
            "description": "Game session ended successfully",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/GameSession"
                }
              }
            }
          },
          "404": {
            "description": "Game session not found",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/GameSession"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/game-sessions": {
      "get": {
        "tags": [
          "Game Sessions"
        ],
        "summary": "Get all game sessions",
        "description": "Retrieves a paginated list of all game sessions",
        "operationId": "getAllGameSessions",
        "parameters": [
          {
            "name": "pageable",
            "in": "query",
            "description": "Pagination parameters",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/Pageable"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Game sessions retrieved successfully",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Page"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Game Sessions"
        ],
        "summary": "Create game session",
        "description": "Creates a new game session with the provided details",
        "operationId": "createGameSession",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GameSessionCreateDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Game session created successfully",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/GameSession"
                }
              }
            }
          },
          "400": {
            "description": "Invalid game session data",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/GameSession"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/game-sessions/{gameSessionName}": {
      "get": {
        "tags": [
          "Game Sessions"
        ],
        "summary": "Get game session by name",
        "description": "Retrieves a game session by its unique name",
        "operationId": "getGameSessionByName",
        "parameters": [
          {
            "name": "gameSessionName",
            "in": "path",
            "description": "Name of the game session to retrieve",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Game session found",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/GameSession"
                }
              }
            }
          },
          "404": {
            "description": "Game session not found",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/GameSession"
                }
              }
            }
          }
        }
      }
    },
    "/api/game-sessions/status/{status}": {
      "get": {
        "tags": [
          "Game Sessions"
        ],
        "summary": "Get game sessions by status",
        "description": "Retrieves a paginated list of game sessions with the specified status",
        "operationId": "getGameSessionsByStatus",
        "parameters": [
          {
            "name": "status",
            "in": "path",
            "description": "Status of the game sessions to retrieve",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "WAITING",
                "IN_PROGRESS",
                "FINISHED"
              ]
            }
          },
          {
            "name": "pageable",
            "in": "query",
            "description": "Pagination parameters",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/Pageable"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Game sessions retrieved successfully",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Page"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/game-sessions/{gameSessionId}": {
      "delete": {
        "tags": [
          "Game Sessions"
        ],
        "summary": "Delete game session",
        "description": "Deletes a game session by its ID",
        "operationId": "deleteGameSession",
        "parameters": [
          {
            "name": "gameSessionId",
            "in": "path",
            "description": "ID of the game session to delete",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "404": {
            "description": "Game session not found"
          },
          "204": {
            "description": "Game session deleted successfully"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "GameSession": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "hostId": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "type": "string"
          },
          "password": {
            "type": "string"
          },
          "maxPlayers": {
            "type": "integer",
            "format": "int32"
          },
          "status": {
            "type": "string",
            "enum": [
              "WAITING",
              "IN_PROGRESS",
              "FINISHED"
            ]
          },
          "startTime": {
            "type": "string",
            "format": "date-time"
          },
          "endTime": {
            "type": "string",
            "format": "date-time"
          },
          "summary": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "players": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/GameSessionPlayer"
            }
          }
        }
      },
      "GameSessionPlayer": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "player": {
            "$ref": "#/components/schemas/User"
          },
          "gameSession": {
            "$ref": "#/components/schemas/GameSession"
          },
          "kills": {
            "type": "integer",
            "format": "int32"
          },
          "deaths": {
            "type": "integer",
            "format": "int32"
          },
          "joinedAt": {
            "type": "string",
            "format": "date-time"
          },
          "leftAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "GrantedAuthority": {
        "type": "object",
        "properties": {
          "authority": {
            "type": "string"
          }
        }
      },
      "User": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "username": {
            "type": "string"
          },
          "passwordHash": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "role": {
            "type": "string",
            "enum": [
              "GUEST",
              "PLAYER",
              "ADMIN"
            ]
          },
          "status": {
            "type": "string",
            "enum": [
              "ONLINE",
              "OFFLINE",
              "IN_GAME",
              "BANNED",
              "INACTIVE"
            ]
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "lastLogin": {
            "type": "string",
            "format": "date-time"
          },
          "userStats": {
            "$ref": "#/components/schemas/UserStats"
          },
          "password": {
            "type": "string"
          },
          "enabled": {
            "type": "boolean"
          },
          "authorities": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/GrantedAuthority"
            }
          },
          "accountNonExpired": {
            "type": "boolean"
          },
          "accountNonLocked": {
            "type": "boolean"
          },
          "credentialsNonExpired": {
            "type": "boolean"
          }
        }
      },
      "UserStats": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "wins": {
            "type": "integer",
            "format": "int32"
          },
          "losses": {
            "type": "integer",
            "format": "int32"
          },
          "kills": {
            "type": "integer",
            "format": "int32"
          },
          "deaths": {
            "type": "integer",
            "format": "int32"
          },
          "matchesPlayed": {
            "type": "integer",
            "format": "int32"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "GameSessionUpdateDTO": {
        "type": "object",
        "properties": {
          "winningTeam": {
            "type": "string"
          },
          "summary": {
            "type": "string"
          }
        },
        "description": "Game session update details"
      },
      "GameSessionCreateDTO": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "hostId": {
            "type": "integer",
            "format": "int64"
          },
          "password": {
            "type": "string"
          }
        },
        "description": "Game session creation details"
      },
...
      "Pageable": {
        "type": "object",
        "properties": {
          "page": {
            "minimum": 0,
            "type": "integer",
            "format": "int32"
          },
          "size": {
            "minimum": 1,
            "type": "integer",
            "format": "int32"
          },
          "sort": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
      "Page": {
        "type": "object",
        "properties": {
          "totalElements": {
            "type": "integer",
            "format": "int64"
          },
          "totalPages": {
            "type": "integer",
            "format": "int32"
          },
          "size": {
            "type": "integer",
            "format": "int32"
          },
          "content": {
            "type": "array",
            "items": {
              "type": "object"
            }
          },
          "number": {
            "type": "integer",
            "format": "int32"
          },
          "sort": {
            "$ref": "#/components/schemas/SortObject"
          },
          "first": {
            "type": "boolean"
          },
          "last": {
            "type": "boolean"
          },
          "numberOfElements": {
            "type": "integer",
            "format": "int32"
          },
          "pageable": {
            "$ref": "#/components/schemas/PageableObject"
          },
          "empty": {
            "type": "boolean"
          }
        }
      },
      "PageableObject": {
        "type": "object",
        "properties": {
          "offset": {
            "type": "integer",
            "format": "int64"
          },
          "sort": {
            "$ref": "#/components/schemas/SortObject"
          },
          "paged": {
            "type": "boolean"
          },
          "pageNumber": {
            "type": "integer",
            "format": "int32"
          },
          "pageSize": {
            "type": "integer",
            "format": "int32"
          },
          "unpaged": {
            "type": "boolean"
          }
        }
      },
      "SortObject": {
        "type": "object",
        "properties": {
          "empty": {
            "type": "boolean"
          },
          "sorted": {
            "type": "boolean"
          },
          "unsorted": {
            "type": "boolean"
          }
        }
      },
...
    }
  }
}